backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.21
mysqlclient==2.2.0
pymysql==1.1.0